requirements.txt
flask>=3.0.0
python-dotenv>=1.0.0
flask-login>=0.6.3
flask-wtf>=1.2.1
werkzeug>=3.0.0
faiss-cpu>=1.13.0
sentence-transformers>=2.2.0
transformers>=4.35.0
torch>=2.9.0
numpy>=1.26.0
python-dateutil>=2.8.0
PyPDF2>=3.0.0
python-docx>=1.0.0
openpyxl>=3.1.0
pandas>=2.0.0
accelerate>=0.24.0
sentencepiece>=0.1.99
authlib>=1.3.0
beautifulsoup4>=4.11.0
feedparser>=6.0.0
selenium>=4.15.0
webdriver-manager>=4.0.0
wikipedia-api>=0.6.0
lxml>=4.9.0
html5lib>=1.1
email-validator>=2.3.0

